pythonmodel/requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
pandas>=2.0.0
scikit-learn>=1.2.0
spacy>=3.5.0
pdfplumber>=0.9.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
